requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
colorama==0.4.6
idna==3.8
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
urllib3==2.2.3
charset-normalizer==3.4.1
Faker==37.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
